cart page added and products data moved to data folder
