+++ src/app/cart/page.tsx
@@ -1,0 +1,19 @@
+export default function Cart() {
+    
+    return (
+      <div className="bg-white">
+        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+          <div className="py-8 md:py-12">
+            <div className="md:grid md:grid-cols-2 md:gap-8">
+              <div className="mt-10 px-4 sm:mt-16 sm:px-0 md:mt-0">
+                <h1 className="text-3xl font-bold tracking-tight text-gray-900">
+                Cart
+                </h1>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  }
+
