--- src/app/products/[id]/loading.tsx
@@ -1,12 +1,0 @@
-export default function Loading() {
-    return (
-        <div className="loading">
-            <div className="loading-gallery" />
-            <div className="loading-info">
-                <div className="loading-title" />
-                <div className="loading-variants" />
-                <div className="loading-button" />
-            </div>
-        </div>
-    );
-}
